pyproject.toml
[tool.poetry]
name = "lab06"
version = "0.1.0"
description = ""
authors = ["damianhealthup <damian@healthup.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "1.23.5"
Pillow = "^9.4.0"
rich = "^13.0.0"
numba = "^0.56.4"

[tool.poetry.group.dev.dependencies]
scipy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
